Add DigiClient model and related test cases
Introduced a new model DigiClient, with fields 'username' and 'password', and a method that allows sending product data to Digi. Tests for this model were also added. In the process, updated the _prepare_barcode method in product_template model to static and modified the imported modules in __init__.py files.