Refine test for sending product data to Digi
The test function for sending product data to Digi has been modified. Instead of simply asserting that the post request was called once, the new approach checks if the correct url and headers were used in this request. This refined test allows for a more detailed verification of the process.