Add product sync and image sync to Digi service
Added two new methods in the product template model to allow the synchronization of product details and images with the Digi service. Whenever a product or its image is updated, it will be automatically sent to Digi. Accompanying tests to ensure this functionality were also added and some code refactoring was performed for better clarity.