Refactor image sending in product_digi_sync
This commit optimizes the process of sending product images in the product_digi_sync module. It moves the condition check for `image_1920` existence inside the `send_image_to_digi` method. Additionally, it transforms the lowercase string of image format from "jpeg" to "jpg" and changes the `LinkNumber` from 1 to 95. Unit tests are also updated to reflect these changes.