Implement create method in product_template and product_category models
Added a create method in the product_template and product_category models to ensure any new records are automatically sent to the digital system. This is done as soon as a new record is created and it has a "plu_code" value for product_template or "external_digi_id" value for product_category.